duel state includes config, determines some behaviours
One goal of this change is to be able to run simulation duels of the computer against itself in order to benchmark various card/deck power. To do that we need a "lobby screen" that the game returns to after all duels are complete, and through which we can determine which decks to use, and whether the player/computer is controlling each side.